refinded the tests and javadoc of TextDatabase and SecureTextDatababse in data package

+++ src/test/java/programming3/chatsys/data/TextDatabaseTest.java
@@ -12,11 +12,14 @@ import java.util.List;
 import java.util.Map;
 
 class TextDatabaseTest {
+
     TextDatabase db;
     File chatMessageDB = new File(".\\message_test.db");
     File userDB = new File(".\\user_test.db");
+
     ChatMessage cm1 = new ChatMessage(1, "Jack_1", new Timestamp(100000), "Haloo");
     ChatMessage cm2 = new ChatMessage(2, "Ana", new Timestamp(200000), "Hello");
+
     User user1 = new User("Jack", "JackMa", "666666");
     User user2 = new User("Jason", "JasonWu", "123456", 1);
     User defaultUser1 = new User("user1\tUser1\tmypassword\t0");
@@ -36,15 +39,31 @@ class TextDatabaseTest {
     @AfterEach
     void tearDown() {
         db = null;
-        chatMessageDB.delete();
-        userDB.delete();
+        if (chatMessageDB.exists())
+            chatMessageDB.delete();
+        if (userDB.exists())
+            userDB.delete();
     }
 
     @Test
     void readMessages() {
-        ChatMessage[] msgArray = new ChatMessage[] {cm1, cm2};
-        List<ChatMessage> msgList = Arrays.asList(msgArray);
+        List<ChatMessage> msgList = Arrays.asList(cm1, cm2);
         assertEquals(msgList, db.readMessages());
+    }
+
+    @Test
+    void testReadMessagesWithGivenNum() {
+        List<ChatMessage> msgList;
+
+        msgList = Arrays.asList(cm2);
+        assertEquals(msgList, db.readMessages(1));
+
+        msgList = Arrays.asList(cm1, cm2);
+        assertEquals(msgList, db.readMessages(2));
+
+        // if request more than the total number of messages in db
+        msgList = Arrays.asList(cm1, cm2);
+        assertEquals(msgList, db.readMessages(1000));
     }
 
     @Test
@@ -59,37 +78,47 @@ class TextDatabaseTest {
 
     @Test
     void addMessage() {
-        ChatMessage cm = new ChatMessage(5, "Jack", new Timestamp(1000000), "Hello World");
-        db.addMessage(cm);
+        db.addMessage(new ChatMessage(5, "Jack", new Timestamp(1000000), "Hello World"));
 
-        ChatMessage[] msgArray = new ChatMessage[] {
+        List<ChatMessage> msgList = Arrays.asList(
                 new ChatMessage(1, "Jack_1", new Timestamp(100000), "Haloo"),
                 new ChatMessage(2, "Ana", new Timestamp(200000), "Hello"),
                 new ChatMessage(5, "Jack", new Timestamp(1000000), "Hello World")
-        };
-        List<ChatMessage> msgList2 = Arrays.asList(msgArray);
+        );
 
-        assertEquals(msgList2, db.readMessages());
+        assertEquals(msgList, db.readMessages());
     }
 
     @Test
-    void getUnreadMessages() {
-        ChatMessage[] msgArray = new ChatMessage[] {
-                new ChatMessage(2, "Ana", new Timestamp(200000), "Hello"),
-        };
-        List<ChatMessage> msgList = Arrays.asList(msgArray);
-        assertEquals(msgList, db.getUnreadMessages("Jason"));
+    void lastId() {
+        int lastId = 0;
+        for (ChatMessage message : db.readMessages()) {
+            if (message.getId() > lastId)
+                lastId = message.getId();
+        }
+        assertEquals(lastId, db.lastId());
+    }
 
-        msgArray = new ChatMessage[] {
+    @Test
+    void testAddMessage() {
+        int lastId = db.lastId();
+        ChatMessage cm = db.addMessage("Jack", "Hello World!!!");
+        assertEquals(cm.getId(), lastId+1);
+
+        List<ChatMessage> msgList = Arrays.asList(
                 new ChatMessage(1, "Jack_1", new Timestamp(100000), "Haloo"),
                 new ChatMessage(2, "Ana", new Timestamp(200000), "Hello"),
-        };
-        msgList = Arrays.asList(msgArray);
-        assertEquals(msgList, db.getUnreadMessages("Jack"));
+                new ChatMessage(cm.getId(), "Jack", cm.getTimestamp(), "Hello World!!!")
+        );
 
-        Map<String, User> userMap = db.readUsers();
-        assertEquals(2, userMap.get("Jack").getLastReadId());
-        assertEquals(2, userMap.get("Jason").getLastReadId());
+        assertEquals(msgList, db.readMessages());
+    }
+
+    @Test
+    void register() {
+        User user = new User("Jane", "JaneDoe", "thisisapassword", 1);
+        db.register(user);
+        assertEquals(user, db.readUsers().get("Jane"));
     }
 
     @Test
@@ -103,4 +132,38 @@ class TextDatabaseTest {
         User user1 = new User("Jack_123", "JackMa", "666666");
         assertTrue(db.register(user1));
     }
+
+    @Test
+    void getUnreadMessages() {
+        List<ChatMessage> msgList = Arrays.asList(
+                new ChatMessage(2, "Ana", new Timestamp(200000), "Hello")
+        );
+        assertEquals(msgList, db.getUnreadMessages("Jason"));
+
+        msgList = Arrays.asList(
+                new ChatMessage(1, "Jack_1", new Timestamp(100000), "Haloo"),
+                new ChatMessage(2, "Ana", new Timestamp(200000), "Hello")
+        );
+        assertEquals(msgList, db.getUnreadMessages("Jack"));
+
+        Map<String, User> userMap = db.readUsers();
+        assertEquals(2, userMap.get("Jack").getLastReadId());
+        assertEquals(2, userMap.get("Jason").getLastReadId());
+    }
+
+    @Test
+    void authenticate() {
+        assertTrue(db.authenticate("Jack", "666666"));
+        assertTrue(db.authenticate("Jason", "123456"));
+        assertFalse(db.authenticate("user1", "password"));
+        assertFalse(db.authenticate("user1", "password"));
+    }
+
+    @Test
+    void getUserIfAuthenticated() {
+        assertEquals(user1, db.getUserIfAuthenticated("Jack", "666666"));
+        assertNull(db.getUserIfAuthenticated("Jason", "666666"));
+        assertNull(db.getUserIfAuthenticated("abc", "666666"));
+    }
+
 }
